Remove infallible variants of functions
