+++ examples/os-strings.rs
@@ -11,25 +11,25 @@ fn main() {
     );
     println!(
         "User's Name            whoami::realname_os():           {:?}",
-        whoami::realname_os(),
+        whoami::realname_os().unwrap_or_else(|_| "<unknown>".to_string().into()),
     );
     println!(
         "User's Username        whoami::username_os():           {:?}",
-        whoami::username_os(),
+        whoami::username_os().unwrap_or_else(|_| "<unknown>".to_string().into()),
     );
     println!(
-        "User's Account         whoami::fallible::account_os():  {:?}",
-        whoami::fallible::account_os()
+        "User's Account         whoami::account_os():            {:?}",
+        whoami::account_os()
             .unwrap_or_else(|_| "<unknown>".to_string().into()),
     );
     println!(
         "Device's Pretty Name   whoami::devicename_os():         {:?}",
-        whoami::devicename_os(),
+        whoami::devicename_os().unwrap_or_else(|_| "<unknown>".to_string().into()),
     );
     println!(
-        "Device's Hostname      whoami::fallible::hostname():    {:?}",
-        whoami::fallible::hostname()
-            .unwrap_or_else(|_| "localhost".to_string()),
+        "Device's Hostname      whoami::hostname():              {:?}",
+        whoami::hostname()
+            .unwrap_or_else(|_| "<unknown>".to_string()),
     );
     println!(
         "Device's Platform      whoami::platform():              {:?}",
